refactor: Changed the folder structure in the project: moved the components and assets folder to the src folder.
