+++ src/app/(app)/[locale]/(client)/page.tsx
@@ -1,4 +1,4 @@
-import Header from './components/header/Header'
+import Header from '../../../../components/header/Header'
 
 export default function Home() {
   return (
