+++ src/components/button/Button.tsx
@@ -1,4 +1,4 @@
-import { IPropsButton } from '../../interfaces/Interfaces'
+import { IPropsButton } from '../interfaces/Interfaces'
 import React, { FC } from 'react'
 
 export const Button: FC<IPropsButton> = ({ className, label, onClick }) => {
