+++ src/components/dropdown/DropDown.tsx
@@ -1,6 +1,6 @@
 'use client'
 import { useState } from 'react'
-import ARROW from '../../../../assets/Arrow.png'
+import ARROW from '@/assets/Arrow.png'
 import Image from 'next/image'
 import Link from 'next/link'
